requirements.txt
openai
streamlit
python-dotenv 
# pytube
pytubefix